move lazy import from utilities to the surgical workflow
